+++ lib/ignoredEvents.js
@@ -1,7 +1,5 @@
 'use strict';
-
-const { name } = require('../package');
-const debug = require('debug')(name);
+const debug = require('debug')('chrome-har');
 
 module.exports = function(method) {
   switch (method) {
