+++ lib/util.js
@@ -1,7 +1,6 @@
 'use strict';
 const urlParser = require('url');
-const { name } = require('../package');
-const debug = require('debug')(name);
+const debug = require('debug')('chrome-har');
 
 const isEmpty = o => !o;
 
@@ -61,5 +60,35 @@ module.exports = {
   },
   isSupportedProtocol(url) {
     return /^https?:/.test(url);
+  },
+  blockedResponse() {
+    return {
+      "status": 0,
+      "statusText": "",
+      "httpVersion": "",
+      "headers": [],
+      "cookies": [],
+      "content": {
+        "size": 0,
+        "mimeType": "x-unknown"
+      },
+      "redirectURL": "",
+      "headersSize": -1,
+      "bodySize": -1,
+      "_transferSize": 0
+    }
+  },
+
+  blockedTimings() {
+    return {
+      blocked: -1,
+      connect: -1,
+      dns: -1,
+      receive: -1,
+      send: -1,
+      ssl: -1,
+      wait: -1,
+      _queued: -1
+    }
   }
 };
